+++ CHANGELOG.md
@@ -9,8 +9,9 @@ Uses save state file version 13.
 ### New features and improvements
 
 - App: Added option to unpause emulator when loading discs. Enabled by default, which changes established behavior.
+- Input: Convert 3D Control Pad analog stick to D-Pad inputs when in digital mode.
 - SH2: Interrupt recalculation microoptimizations.
-- VDP1 software renderer performance microoptimizations:
+- VDP1: Software renderer performance microoptimizations:
     - Do these once per command instead of per pixel:
         - Determine double density mode
         - Get references to VDP1 registers